+++ .eslintrc.js
@@ -1,3 +1,12 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires, import/no-extraneous-dependencies
+const a11yOff = Object.keys(require('eslint-plugin-jsx-a11y').rules).reduce(
+  (acc, rule) => {
+    acc[`jsx-a11y/${rule}`] = 'off';
+    return acc;
+  },
+  {},
+);
+
 module.exports = {
   root: true,
   env: {
@@ -9,9 +18,10 @@ module.exports = {
   extends: [
     // 'eslint:recommended',
     'plugin:@typescript-eslint/recommended',
-    'airbnb-base',
-    'prettier',
+    'airbnb',
+    'prettier'
   ],
+
   settings: {
     'import/resolver': {
       node: {
@@ -20,6 +30,9 @@ module.exports = {
     },
   },
   rules: {
+    'no-use-before-define': 'off',
+    '@typescript-eslint/no-use-before-define': ['error'],
+
     '@typescript-eslint/explicit-module-boundary-types': 0,
     'import/no-unresolved': [2, { ignore: ['@'] }],
     'import/prefer-default-export': 0,
@@ -34,5 +47,13 @@ module.exports = {
         tsx: 'never',
       },
     ],
+
+    // React
+    'react/jsx-filename-extension': [2, { extensions: ['.tsx'] }],
+    'react/jsx-props-no-spreading': 0,
+    'react/react-in-jsx-scope': 0,
+
+    // JSX a11y
+    ...a11yOff,
   },
 };