+++ pages/traits/all.tsx
@@ -1,32 +1,35 @@
-import { traits } from '@/traits';
+import HtmlContent from '@/components/HtmlContent';
+import { Trait, traits } from '@/traits';
 import { useTranslation } from '@/translation/useTranslation';
 
 const Traits = () => {
   const t = useTranslation();
 
-  const allTraits = traits
-    .sort(
-      (
-        { id: idA, deprecated: aDeprecated },
-        { id: idB, deprecated: bDeprecated },
-      ) => {
-        if (aDeprecated && !bDeprecated) return 1;
-        if (!aDeprecated && bDeprecated) return -1;
+  const allTraits = (traits as (Trait<'weighted'> & Trait<'deprecated'>)[]).sort(
+    (
+      { id: idA, deprecated: aDeprecated },
+      { id: idB, deprecated: bDeprecated },
+    ) => {
+      if (aDeprecated && !bDeprecated) return 1;
+      if (!aDeprecated && bDeprecated) return -1;
 
-        if (idA < idB) return -1;
-        if (idA > idB) return 1;
+      if (idA < idB) return -1;
+      if (idA > idB) return 1;
 
-        return 0;
-      },
-    );
+      return 0;
+    },
+  );
 
   return (
     <>
       <div className="section">
         <div className="container is-max-widescreen">
           <div className="columns is-multiline">
-            {allTraits.map((trait) => (
-              <div key={trait.id} className="column is-12 is-6-desktop is-4-widescreen">
+            {allTraits.map(trait => (
+              <div
+                key={trait.id}
+                className="column is-12 is-6-desktop is-4-widescreen"
+              >
                 <div
                   className={[
                     'card',
@@ -40,11 +43,9 @@ const Traits = () => {
                     </div>
                   </div>
                   <div className="card-content">
-                    <div
+                    <HtmlContent
                       className="content px-2 pb-2"
-                      dangerouslySetInnerHTML={{
-                        __html: t(`traits:${trait.id}.description`),
-                      }}
+                      html={t(`traits:${trait.id}.description`)}
                     />
                   </div>
                 </div>