+++ src/components/HtmlContent.tsx
@@ -1,0 +1,10 @@
+/* eslint-disable react/no-danger */
+type Props = {
+  html: string,
+} & { [key: string]: unknown };
+
+const HtmlContent = ({ html, ...props }: Props) => (
+  <div {...props} dangerouslySetInnerHTML={{ __html: html }} />
+);
+
+export default HtmlContent;